Split model and data to separate files for bertmesh
